@@ -1,11 +1,10 @@
 module server
 
-go 1.23.4
+go 1.24.0
 
 require (
+	github.com/go-chi/chi v1.5.5
 	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
 	gopkg.in/yaml.v3 v3.0.1
 )
-
-require github.com/go-chi/chi v1.5.5 // indirect
