@@ -1,11 +1,17 @@
 package server
 
 import (
+	"encoding/json"
 	"fmt"
 	"log"
+	"log/slog"
 	"net/http"
+	"os"
+	"path/filepath"
+
 	"server/config"
 	"server/db"
+	"server/models"
 
 	"github.com/go-chi/chi"
 )
@@ -15,27 +21,51 @@ type Server struct {
 }
 
 func Run(cfg config.Config) error {
+	// этот кусок
+	logDir, logFile := "logger", "sysLog.log"
+	logPath := filepath.Join(logDir, logFile)
+
+	file, err := os.OpenFile(logPath, os.O_CREATE|os.O_WRONLY|os.O_TRUNC, 0666)
+	if err != nil {
+		slog.Error("opening log file", "error", err)
+		return err
+	}
+	defer file.Close()
+
+	opts := &slog.HandlerOptions{
+		AddSource: true,
+		Level:     slog.LevelDebug,
+	}
+
+	Logger := slog.New(slog.NewJSONHandler(file, opts))
+
+	slog.SetDefault(Logger)
+	// до этого момента
+
 	router := chi.NewRouter()
 
-	s, err := NewServer()
+	s, err := newServer()
 	if err != nil {
 		log.Println(err)
 		return err
 	}
 
-	fmt.Println(s)
+	router.Get("/admin/info", s.InfoHandler)
+	router.Post("/payOrder", s.PayOrder)
 
-	// router.Get("/admin/info", s.InfoHandler)
-
-	log.Printf("Starting HTTP server on port %d", cfg.Port)
+	slog.Info("starting HTTP server on port")
 	if err := http.ListenAndServe(fmt.Sprintf(":%d", cfg.Port), router); err != nil {
-		log.Fatalf("Error starting HTTP server: %v", err)
+		slog.Info("starting HTTP server error")
 	}
+
+	// data := bl.OrderManage(s.Db)
+
+	// fmt.Println(data)
 
 	return nil
 }
 
-func NewDB() (*db.Database, error) {
+func newDB() (*db.Database, error) {
 	dbParams, err := config.GetDBParams()
 	if err != nil {
 		log.Println(err)
@@ -51,10 +81,11 @@ func NewDB() (*db.Database, error) {
 	return database, nil
 }
 
-func NewServer() (*Server, error) {
-	database, err := NewDB()
+func newServer() (*Server, error) {
+
+	database, err := newDB()
 	if err != nil {
-		log.Println(err)
+		slog.String("error", err.Error())
 		return nil, err
 	}
 
@@ -63,4 +94,21 @@ func NewServer() (*Server, error) {
 	}
 
 	return s, nil
+}
+
+func (s *Server) InfoHandler(w http.ResponseWriter, r *http.Request) {
+	w.Write([]byte("Server info"))
+}
+
+func (s *Server) PayOrder(w http.ResponseWriter, r *http.Request) {
+	var payStatus models.Payment
+	if err := json.NewDecoder(r.Body).Decode(&payStatus); err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		json.NewEncoder(w).Encode(err)
+		return
+	}
+	defer r.Body.Close()
+
+	order := models.Order{}
+	log.Printf("new factor %v", order.Payment)
 }
